Fix warnings on NetBSD
The isdigi() and isxdigit() functions take 'int' argument, older GCC
will complain about passing chars to them.

+++ src/cprintf.c
@@ -51,7 +51,7 @@ int cprintf(const char *fmt, ...)
 	while (*fmt) {
 		if (*fmt == '%') {
 			num = ptr = (char *)&fmt[1];
-			while (*ptr && isdigit(*ptr))
+			while (*ptr && isdigit((int)*ptr))
 				ptr++;
 
 			if (*ptr == 'C') {
